+++ custom_components/beurer_daylight_lamps/const.py
@@ -6,7 +6,7 @@ import logging
 from typing import Final
 
 DOMAIN: Final = "beurer_daylight_lamps"
-VERSION: Final = "1.32.0"
+VERSION: Final = "1.33.0"
 LOGGER = logging.getLogger(__package__)
 
 # BLE Characteristic UUIDs